changes on task container
